Preparing reporting for submodules analysis
